requirements.txt
awscli==1.16.262
bcrypt==3.1.7
certifi==2019.9.11
colorama==0.4.1
Django==2.2.7
docutils==0.15.2
jmespath==0.9.4
pyasn1==0.4.7
python-dateutil==2.8.0
pytz==2019.3
PyJWT==1.7.1
PyYAML==5.1.2
rsa==3.4.2
s3transfer==0.2.1
SQLAlchemy==1.3.11
sqlparse==0.3.0
django-cors-headers==3.2.0
mysqlclient==1.4.6
requests==2.22.0